♻️ refactor: migrate phase-optional fields to parseOptional (#130 part 2)
Replace `as? T` silent-coerce on phase-level optional fields (prompt,
template, source, exclude_self, options, rounds, if) with `parseOptional<T>`.
A YAML typo like `rounds: "3"` (accidentally quoted) now throws instead of
silently nil-ing out and defaulting to 1.

Split tests into `ScenarioLoaderTests+StrictTypes.swift` extension per
`.claude/rules/testing.md` (extension, not a new `@Suite`, to keep
`.serialized`/`.timeLimit` inherited). Drop `private` from the two helpers
called across the split (`makeYAMLWithAssignTarget`, `makeMinimalYAML(phasesBlock:)`).

Negative cases: quoted rounds/exclude_self, int exclude_self, mixed-type options.
Positive case: `exclude_self: yes` stays valid — YAML 1.1 boolean, pinned so
a future "fix" doesn't break it.
